+++ lib/js_regex/conversion.rb
@@ -13,25 +13,27 @@ class JsRegex
     require_relative 'second_pass'
 
     class << self
-      def of(ruby_regex, options: nil)
-        source, warnings = convert_source(ruby_regex)
-        options_string   = convert_options(ruby_regex, options)
+      def of(input, options: nil)
+        source, warnings = convert_source(input)
+        options_string   = convert_options(input, options)
         [source, options_string, warnings]
       end
 
       private
 
-      def convert_source(ruby_regex)
-        tree = Regexp::Parser.parse(ruby_regex)
+      def convert_source(input)
+        tree = Regexp::Parser.parse(input)
         context = Converter::Context.new(case_insensitive_root: tree.i?)
         converted_tree = Converter.convert(tree, context)
         final_tree = SecondPass.call(converted_tree)
         [final_tree.to_s, context.warnings]
       end
 
-      def convert_options(ruby_regex, custom_options)
+      def convert_options(input, custom_options)
         options = custom_options.to_s.scan(/[gimuy]/)
-        options << 'i' if (ruby_regex.options & Regexp::IGNORECASE).nonzero?
+        if input.is_a?(Regexp) && (input.options & Regexp::IGNORECASE).nonzero?
+          options << 'i'
+        end
         options.uniq.sort.join
       end
     end